# api.yaml
openapi: 3.0.3
info:
  title: Macro
  description: Все табличные страницу имеют схожую структуру Schemas - table
  version: 1.0.0
servers:
  - url: 'http://127.0.0.1:4010'
paths:
  /api/audiences:
    get:
      summary: Получение аудиторий
      description: Возвращает все аудитории
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Audience'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudienceCreateRequest'
      responses:
        201:
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audience'
  /api/audiences/{audienceId}:
    delete:
      description: Удалить аудитория
      parameters:
        - name: audienceId
          in: path
          schema:
            type: integer
          example: 1
          required: true
      responses:
        204:
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
  /api/audiences/{audienceId}/disconnect:
    delete:
      description: Удалить все подключения
      parameters:
        - name: audienceId
          in: path
          schema:
            type: integer
          example: 1
          required: true
      responses:
        204:
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
  /api/audiences/integrations:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationRequest'
      responses:
        201:
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    Cabinet:
      type: object
      description: Рекламный кабинет
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          description: yandex, google, facebook
          example: yandex
    Integration:
      type: object
      description: Вид подключения для отображения
      properties:
        cabinet_name:
          type: string
          example: yandex
#          $ref: '#/components/schemas/Cabinet'
        id:
          type: integer
          description: Идентификатор интеграции
          example: 1
    IntegrationRequest:
      type: object
      description: Запрос на подключение
      properties:
        cabinet_name:
          type: string
          example: yandex
          description: название кабинета
        audience_ids:
          type: array
          items:
            type: integer
            example: 1
          example: [1, 2, 3]
    Audience:
      type: object
      description: Аудитория
      properties:
        id:
          type: integer
          description: Идентификатор аудитории
          example: 1
        name:
          type: string
          description: Название
          example: "Бронь"
        duration:
          type: string
          description: Продолжительность
          example: '1 день 3 часа'
        created_at:
          type: string
          description: Дата создания
          example: '2024-11-14T21:00:00.000Z'
        updated_at:
          type: string
          description: Дата изменения
          example: '2024-11-14T21:00:00.000Z'
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/Integration'
    AudienceCreateRequest:
      type: object
      description: необходимо выбрать или список заявок или критерии отбора
      properties:
        name:
          type: string
          description: Название
          example: "Бронь"
          #required: true
        filter:
          $ref: '#/components/schemas/AudienceCreateRequestFilter'
#          required: true
#          type: object
    AudienceCreateRequestFilter:
      type: object
      minProperties: 1
      properties:
        creation_date_from:
          type: string
          example: Sat Oct 19 2024 17:47:50 GMT+0300 (Москва, стандартное время)
        creation_date_to:
          type: string
          example: Sat Oct 19 2024 17:47:50 GMT+0300 (Москва, стандартное время)
        statuses:
          type: array
          items:
            type: string
        rejection_reasons:
          type: array
          items:
            type: string
        non_target_reasons:
          type: array
          items:
            type: string
        applications:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
...